Fix warnings on x86-64
git-svn-id: https://piggledy.org/svn/peerfuse/trunk@739 f7b84f85-2e04-4de5-9e99-c4612dc0539c

+++ common/file_content.cpp
@@ -100,7 +100,7 @@ void FileContent::Truncate(off_t offset)
 {
 	BlockLockMutex lock(this);
 
-	if(offset > file_size)
+	if(offset > (off_t)file_size)
 		return;
 
 	iterator it = begin();
